CLN: #TW-44 webDto 클래스명 재정립
