+++ src/main/java/com/interior/abstraction/domain/IQuery.java
@@ -1,0 +1,5 @@
+package com.interior.abstraction.domain;
+
+public interface IQuery {
+
+}
