+++ src/main/java/com/interior/adapter/inbound/business/webdto/CreateBusinessMaterialWebDtoV1.java
@@ -1,0 +1,19 @@
+package com.interior.adapter.inbound.business.webdto;
+
+import java.math.BigDecimal;
+
+public class CreateBusinessMaterialWebDtoV1 {
+
+    public record CreateBusinessMaterialReqDto(
+            String materialName,
+            String materialUsageCategory,
+            String materialCategory,
+            BigDecimal materialAmount,
+            String materialAmountUnit,
+            String materialMemo,
+            String materialCostPerUnit,
+            String laborCostPerUnit
+    ) {
+
+    }
+}
