+++ src/main/java/com/interior/adapter/inbound/business/webdto/GetBusinessMaterialLogsWebDtoV1.java
@@ -1,9 +1,9 @@
-package com.interior.application.readmodel.business.dto;
+package com.interior.adapter.inbound.business.webdto;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import java.time.LocalDateTime;
 
-public class GetBusinessMaterialLogs {
+public class GetBusinessMaterialLogsWebDtoV1 {
 
     public record Res(
             String updaterName,
