--- src/main/java/com/interior/adapter/inbound/business/webdto/GetExcelOfBusinessMaterialListWebDtoV1.java
@@ -1,9 +1,0 @@
-package com.interior.adapter.inbound.business.webdto;
-
-public class GetExcelOfBusinessMaterialListWebDtoV1 {
-
-    public record Res(String taskId) {
-
-    }
-
-}
