--- src/main/java/com/interior/adapter/inbound/business/webdto/ReviseBusiness.java
@@ -1,7 +1,0 @@
-package com.interior.adapter.inbound.business.webdto;
-
-public class ReviseBusiness {
-
-    public record WebReqV1(String changeBusinessName) { }
-
-}
