+++ src/main/java/com/interior/adapter/inbound/business/webdto/ReviseBusinessWebDtoV1.java
@@ -1,0 +1,9 @@
+package com.interior.adapter.inbound.business.webdto;
+
+public class ReviseBusinessWebDtoV1 {
+
+    public record Req(String changeBusinessName) {
+
+    }
+
+}
