--- src/main/java/com/interior/adapter/inbound/business/webdto/ReviseUsageCategoryOfMaterial.java
@@ -1,10 +1,0 @@
-package com.interior.adapter.inbound.business.webdto;
-
-import java.util.List;
-
-public class ReviseUsageCategoryOfMaterial {
-
-    public record Req(List<Long> subDataIds, String usageCategoryName) { }
-
-    public record Res(String changedUsageCategoryName) { }
-}
