+++ src/main/java/com/interior/adapter/inbound/business/webdto/ReviseUsageCategoryOfMaterialWebDtoV1.java
@@ -1,0 +1,14 @@
+package com.interior.adapter.inbound.business.webdto;
+
+import java.util.List;
+
+public class ReviseUsageCategoryOfMaterialWebDtoV1 {
+
+    public record Req(List<Long> subDataIds, String usageCategoryName) {
+
+    }
+
+    public record Res(String changedUsageCategoryName) {
+
+    }
+}
