+++ src/main/java/com/interior/adapter/inbound/business/webdto/SendQuotationDraftToClientWebDtoV1.java
@@ -1,8 +1,9 @@
 package com.interior.adapter.inbound.business.webdto;
 
-public class SendQuotationDraftToClient {
+public class SendQuotationDraftToClientWebDtoV1 {
 
     public record Req(String receiverPhoneNumber) {
+
     }
 
 }
